use cucumber and mockito to do the unit test
