+++ Utils/azuretools-core/test/resources/com/microsoft/azuretools/azurecommons/rediscacheprocessors/RedisCacheCreatorScenario.feature
@@ -1,0 +1,4 @@
+Feature: Test RedisCacheCreator
+  Scenario: Initialize the RedisCacheCreator
+    Given initialize the RedisCacheCreator
+    Then check the the key number of CreatorMap
